+++ packages/fraxel/src/nodes/node2d/lib/drop-area-system.ts
@@ -116,6 +116,8 @@ export class DropAreaSystem {
   }
 
   static #visitInDrawOrder(node: Node, cb: (node: Node) => void) {
+    if (!node.active) return
+
     cb(node)
     for (const child of node._children) {
       this.#visitInDrawOrder(child, cb)